+++ packages/bloc_annotation/lib/src/bloc.dart
@@ -1,0 +1,13 @@
+import 'package:bloc_annotation/src/annotation.dart';
+
+final class BlocMeta extends BaseAnnotation {
+  const BlocMeta({
+    super.name,
+    super.copyWith,
+    super.overrideToString,
+    super.overrideEquality,
+    this.state,
+  });
+
+  final Type? state;
+}