+++ packages/bloc_annotation/lib/src/state.dart
@@ -1,0 +1,13 @@
+import 'package:bloc_annotation/src/annotation.dart';
+
+final class StateMeta extends BaseAnnotation {
+  const StateMeta({
+    super.name,
+    super.copyWith,
+    super.overrideToString,
+    super.overrideEquality,
+    this.isSealed = true,
+  });
+
+  final bool isSealed;
+}