+++ packages/bloc_annotation/CHANGELOG.md
@@ -1,3 +1,9 @@
-## 1.0.0
+# Changelog
 
-- Initial version.
+## [1.0.0] - 2025-11-24
+
+### Added
+- Added `@Bloc` annotation for defining BLoC classes.
+- Added `@Cubit` annotation for defining Cubit classes.
+- Added `@State` annotation for defining state classes or methods.
+- Added `@Event` annotation for defining event methods.